+++ src/models.rs
@@ -175,6 +175,17 @@ impl Display for Category {
     }
 }
 
+impl Display for Status {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        match self {
+            Status::Created => write!(f, "CREATED"),
+            Status::Highlighted => write!(f, "HIGHLIGHTED"),
+            Status::Completed => write!(f, "COMPLETED"),
+            Status::Archived => write!(f, "ARCHIVED"),
+        }
+    }
+}
+
 #[cfg(test)]
 mod tests {
     use super::initials;